Created asset folder then CSS folder and moved my style.css file
Added a new folder named 'assets' to the main directory. and another folder by the name of css, This folder will be used to store various assets such as images, stylesheets, and other resources needed for the project. The structure is set up to improve organization and ensure a clean separation of code and assets.